Break up `AnalyzeProperties` into smaller modules.
